FINDING: LV-SAFETY-DOT-INSPECTION-CREATE-DUAL-TYPE-42P08 LANE: NON-FINANCIAL
ROOT CAUSE: Post-#8892 Live retry reached the corrected dual-column INSERT but Render request req-65 proved PostgreSQL 42P08: one uncast $7 parameter was inferred as both legacy inspection_level SMALLINT and canonical fmcsa_level INTEGER.

FIX: Explicitly cast the one validated level as smallint and integer at the two compatibility targets. Tightened the focused test and already-wired planted-defect guard to require the exact casts.

DOD-A: PASS
DOD-B: N/A
DOD-C: PASS
DOD-D: N/A
DOD-E: PASS
VERIFY-1: PASS
VERIFY-2: PASS
VERIFY-3: PASS
VERIFY-4: PASS
VERIFY-5: N/A
VERIFY-6: N/A
VERIFY-7: PASS
VERIFY-8: PASS
MODULE_PROGRESS: safety 38 of 38 Built; Live remains blocked pending deploy recheck
ITEMS_TOUCHED: LV-SAFETY-DOT-INSPECTION-CREATE-DUAL-TYPE-42P08
MIGRATE: N/A

GUARD: scripts/verify-safety-dot-inspection-unit-entitypicker.mjs + scripts/verify-steps/2164-verify-safety-dot-inspection-unit-entitypicker.mjs
LIVE PROOF: Render request req-65 at 2026-08-18T11:53:15Z returned pg_code 42P08 with detail smallint versus integer; guard normal and --selftest exit 0; focused route tests 3/3; root tsc exit 0; audit scoreboard exit 0.
REMAINING: Deploy and repeat selected-USMCA create/reload/drill before Live credit. Product Live=BLOCKED.

+++ apps/backend/src/routes/safety/dot-inspections-csa-source.routes.test.ts
@@ -99,7 +99,7 @@ describe("DOT inspection CSA canonical rollup", () => {
     const sqlText = mockQuery.mock.calls.map((call) => String(call[0])).join("\n");
     expect(sqlText).not.toMatch(/score_date|source_dot_inspection_count/);
     expect(sqlText).toMatch(/inspection_level, fmcsa_level/);
-    expect(sqlText).toMatch(/\$7,\$7/);
+    expect(sqlText).toMatch(/\$7::smallint,\$7::integer/);
   });
 
   it("persists an exact per-category breakdown for multi-category inspections", async () => {
